fix(daemon): send polkit uid as INT32 not UINT32
Polkit's CheckAuthorization D-Bus method expects the 'uid' field in
the unix-process subject dict as INT32 (D-Bus type 'i'), but we were
sending it as UINT32 (D-Bus type 'u'). Polkit's lookup_asv() does a
strict GVariant type match and silently skipped the uid value, leaving
it as the default -1, which produced the error:

  'Unix process subject does not have uid set'

This caused the GUI unlock to fail with 'privileged operations
require root access (polkit unavailable)' even though polkit was
running and the action policy was installed correctly.

Fix: cast uid from u32 to i32 before inserting into the subject dict.

+++ crates/daemon/src/dbus/auth.rs
@@ -128,7 +128,9 @@ async fn check_polkit_authorization(uid: u32, pid: u32) -> Result<bool, String> 
     let subject_dict: HashMap<&str, Value<'_>> = {
         let mut m = HashMap::new();
         m.insert("pid", Value::from(pid));
-        m.insert("uid", Value::from(uid));
+        // Polkit expects uid as INT32 (D-Bus type 'i'), not UINT32.
+        // See polkitsubject.c: lookup_asv(details, "uid", G_VARIANT_TYPE_INT32, ...).
+        m.insert("uid", Value::from(uid as i32));
         m.insert(
             "start-time",
             Value::from(get_process_start_time(pid).unwrap_or(0u64)),
